fix v2

+++ diff-pairs/diff_pairs.h
@@ -14,7 +14,7 @@ int64_t CountPairs(const std::vector<int>& data, int x) {
     int64_t res = 0;
     int64_t search_value;
     for (int i = 0; i < std::ssize(data); ++i) {
-        search_value = x - data[i];
+        search_value = static_cast<int64_t>(x) - data[i];
         if (positions.contains(search_value)) {
             for (int p : positions[search_value]) {
                 if (p > i) {
